requirement.txt
streamlit
ollama
langchain
langchain-core
langchain-ollama
python-dotenv
